Fix 'make distcheck'
In particular this snippet

 giomodulesdir=`pkg-config --variable=giomoduledir gio-2.0`

is wrong as it breaks 'make distcheck' unless you can write to where
the gio modules are installed. While this is true for jhbuild setups
it is patently not true for normal setups.

In fact, one could argue that glib's --with-gio-module-dir option is
incompatible with how we build things, but that is more of a GLib
problem.